\name{BayesTestStreak-package}
\alias{BayesTestStreak-package}
\alias{BayesTestStreak}
\docType{package}
\title{
Testing Streakiness in Binary Sequences by Bayes Factors
}
\description{
BayesTestStreak contains a collection of functions for comparing consistent and streaky models for a binary sequence using Bayes factors.  The methods are based on the group of spacings between consecutive successes.
}
\details{
\tabular{ll}{
Package: \tab BayesTestStreak\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-04-08\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Jim Albert

Maintainer: Jim Albert <albert@bgsu.edu>
}
\references{
Albert, J. (2013), Looking at Spacings to Access Streakiness, Journal of Quantitative Analysis of Sports, Vol 9, Issue 2, 151-163.
}
\keyword{ package }

