\name{dustin_streak}
\alias{dustin_streak}
\docType{data}
\title{Streak Data for Dustin Pedroia}
\description{
Hit/out data for Dustin Pedroia for his career through August 2016
}
\usage{
dustin_streak
}
\format{
  A data frame on the following two variables.
  \describe{
  \item{Season}{season}
  \item{Outcome}{binary outcome if hit (1) or out (0)}
}
}
\source{Retrosheet data}
\keyword{datasets}
