\name{streak_data}
\alias{find_id}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Finds Retrosheet id for a player
}
\description{
Given a player's name, finds the Retrosheet id code.
}
\usage{
find_id(player)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{player}{
Name of the player (character type)
}
}
\value{
Retrosheet id
}
\author{
Jim Albert
}

\examples{
find_id("Roger Maris")
}

